Moving NewsfeedService construction into its own class and storing it on MainActivity so that it's only constructed once.
