+++ app/src/main/res/layout/newsfeed_item_view.xml
@@ -9,7 +9,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="8dp"
-        android:layout_marginBottom="8dp"
+        android:paddingBottom="8dp"
         android:textSize="16sp"
         app:layout_constraintBottom_toTopOf="@id/newsfeed_item_title"
         app:layout_constraintEnd_toEndOf="parent"
